LRCI-6699 Add ability to use an OS Architecture specific label
